testsuite: fixlets and improvements in two long TAP-based tests
* tests/depmod.tap: Clean up the subdirectories used by tests that
passed, to avoid ending up with a too big test directory.  This is
especially important since, on each and every system, at least some
tests in this script are expected to be skipped (which ones exactly
depends on the system), thus causing the test directory not to be
removed when the script terminates.
* tests/instspc.tap: Likewise (with the difference that tests here
are not expected to be skipped, but to xfail).  Give more explicit
error message in case of setup failure.  Skip the proper number of
tests when required, to avoid spurious errors due to mismatched
plan.  Also, really avoid to run the unsupported test cases when
the need to skip them is detected.