automake: refactor, break up 'require_file_internal'
This refactoring is only required in view of future changes.

* automake.in (require_file_internal): Move the guts of this
function ...
(required_file_check_or_copy): ... into this new function.  This
ensures that calls to `push_required_file' and code that copies
required files are placed in separate functions; this will be
very useful for reorganizing de-serialization of file installs
in future changes.