tests: fix spurious failures w.r.t. parallel make and colorization
* tests/color2.test: Skip the test if the $MAKE program fails to
consider the standard output as a tty when spawned by `expect'.
This is required for make implementations, like FreeBSD make and
Solaris dmake, that redirect the output of recipes to temporary
files or pipes when run in parallel mode.  Since we are at it,
simplify the detection of a working `expect' program, and throw
in other minor simplifications.