tests: various minor tweakings, mostly related to AM_PROG_AR
* tests/alloca2.test: Ensure we don't experience a spurious failure
due to a missing `AM_PROG_AR' macro or a missing `ar-lib' auxiliary
script.
* tests/libtool4.test: Likewise.
* tests/ldadd.test: Likewise.  Since we are at it, make grepping of
automake stderr stricter.
* tests/reqd2.test: Likewise.
* tests/pr211.test: Ensure automake fails also with `-Wnone', since
the error we are testing for is an hard error, not a mere warning.
* tests/syntax.test: Likewise, and ensure we don't fail to other
errors by removing use of `lib_LTLIBRARIES' in Makefile.am.  Since
we are at it, make grepping of automake stderr stricter.