tests: fix spurious failure in 'color2.test'
* tests/color2.test: (expect-make): Add an "expect eof" directive,
so that the collected output from the spawned make program will be
displayed on stdout, as desired.
Since we are at it, also correctly remove a temporary file which
we was trying to delete using a wrong filename.

Problem introduced in merge `v1.11-1579-g8d3466c', probably by a
botched edit or conflict resolution.