tests: remove FIXME comments from tricks to pacify syntax checks
Originally, I myself had insisted on adding those "FIXME" comments
near every unusual construct or use of "creative quoting" whose
only purpose was to avoid false positive from out syntax checks.
I did that with the expectation that we could have easily and soon
added a whitelisting capability to our maintainer-specific syntax
checks.  This hasn't been the case unfortunately, and now those
"FIXME" in the comments are just distracting noise, making it more
difficult to grep the test cases for stuff that could really use a
fix or an improvement.  So just get rid of them.

* tests/instdir-ltlib.test: Remove that extra "FIXME".
* tests/instdir-prog.test: Likewise.
* tests/parallel-tests3.test: Likewise.
* tests/parallel-tests-dry-run.test: Likewise.
* tests/parallel-tests-console-output.test: Likewise.
* tests/parallel-tests-no-color-in-log.test: Likewise.
* tests/tap-realtime.test: Likewise.
* tests/test-trs-recover.test: Likewise.
* tests/tap-deps.test: Likewise.