docs: fix a couple broken anchors
* doc/automake.texi: In previous commit v1.11-605-g5f71f2b (related
to automake bug#8071), we had renamed the chapters about aclocal
and automake invocation, and added anchors to the old names as not
to break hyperlinks to our on-line manual.  Unfortunately, these
anchors had been erroneously placed *before* the relevant `@node'
commands, instead that just after, so that they ended up pointing
to the wrong chapters.  Fix this.  Since we are at it, tweak the
wording of the comments to the anchors to better match that used
in the autoconf manual.

Report and suggestion by Eric Blake.